+++ src/tests/cairo_run_test.rs
@@ -1311,6 +1311,13 @@ fn cairo_run_uint384_extension() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_field_arithmetic() {
+    let program_data = include_bytes!("../../cairo_programs/field_arithmetic.json");
+    run_program_simple(program_data.as_slice());
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn cairo_run_ed25519_field() {
     let program_data = include_bytes!("../../cairo_programs/ed25519_field.json");
     run_program_simple(program_data.as_slice());
@@ -1339,6 +1346,13 @@ fn cairo_run_secp256r1_slope() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_run_garaga() {
+    let program_data = include_bytes!("../../cairo_programs/garaga.json");
+    run_program_simple(program_data.as_slice());
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn cairo_run_div_mod_n() {
     let program_data = include_bytes!("../../cairo_programs/div_mod_n.json");
     run_program_simple(program_data.as_slice());
@@ -1360,7 +1374,21 @@ fn cairo_run_is_zero_pack() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-fn cairo_run_quad_bit() {
-    let program_data = include_bytes!("../../cairo_programs/quad_bit.json");
+fn cairo_run_n_bit() {
+    let program_data = include_bytes!("../../cairo_programs/n_bit.json");
+    run_program_simple(program_data.as_slice());
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_run_highest_bitlen() {
+    let program_data = include_bytes!("../../cairo_programs/highest_bitlen.json");
+    run_program_simple(program_data.as_slice());
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_run_uint256_improvements() {
+    let program_data = include_bytes!("../../cairo_programs/uint256_improvements.json");
     run_program_simple(program_data.as_slice());
 }